+++ tests/macro_optimised.rs
@@ -22,7 +22,7 @@ struct SimpleOptimised {
 	b: u32,
 }
 
-#[revisioned(revision(1, encoding = "optimised", struct = "indexed"))]
+#[revisioned(revision(1, encoding = "optimised", indexed_struct))]
 #[derive(Debug, Clone, PartialEq)]
 struct IndexedOptimised {
 	a: u32,